feat: add favicon and per-route page titles (#187)
- Add public/favicon.svg with a simple robot/fleet icon
- Update layout.tsx metadata with title template "%s | Fleet Dashboard"
- Extract OverviewContent.tsx as client component from page.tsx
- Convert page.tsx and all route pages to server components to enable
  metadata exports with per-route titles (Overview, Agents, Pull Requests,
  Services, Costs, Settings)
